Data Scrape Stream/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
albucore==0.0.13
albumentations==1.4.10
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
asn1crypto==1.5.1
astor==0.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.49
botocore==1.35.49
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cryptography==43.0.3
cssselect==1.2.0
cycler==0.12.1
Cython==3.0.11
dataclasses-json==0.6.7
decorator==5.1.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
environs==9.5.0
et_xmlfile==2.0.0
faiss-cpu==1.9.0
fake-useragent==1.5.1
fastapi==0.115.4
filelock==3.16.1
fire==0.7.0
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.1
idna==3.10
imageio==2.36.0
imgaug==0.4.0
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.43
langchain-nvidia-ai-endpoints==0.1.2
langchain-text-splitters==0.2.1
langsmith==0.1.139
lazy_loader==0.4
llama-index-core==0.11.19
llama-index-embeddings-nvidia==0.2.3
llama-index-embeddings-openai==0.2.5
llama-index-llms-nvidia==0.2.6
llama-index-llms-openai==0.2.16
llama-index-llms-openai-like==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-opendal==0.0.1
llama-index-vector-stores-milvus==0.2.7
llama-index-vector-stores-pinecone==0.2.1
llama-parse==0.5.11
lmdb==1.5.1
lxml==5.3.0
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
matplotlib==3.9.2
mdurl==0.1.2
milvus-lite==2.4.10
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.12.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
openai==1.52.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.10.10
packaging==24.1
paddleocr==2.9.1
paddlepaddle==2.6.2
pandas==2.2.3
parse==1.20.2
pdf2image==1.17.0
pillow==10.4.0
pinecone==5.1.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
propcache==0.2.0
protobuf==4.25.5
protoc-gen-openapiv2==0.0.1
pyarrow==18.0.0
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pydeck==0.9.1
pyee==12.0.0
Pygments==2.18.0
PyJWT==2.9.0
pymilvus==2.4.8
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyOpenSSL==24.2.1
pyparsing==3.2.0
pypdf==4.3.1
pyppeteer==0.0.25
pyquery==2.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-html==0.10.0
requests-toolbelt==1.0.0
rich==13.9.3
rpds-py==0.20.0
s3transfer==0.10.3
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.2.0
shapely==2.0.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowflake-connector-python==3.12.3
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.2
streamlit==1.39.0
streamlit-option-menu==0.4.0
streamlit-pdf-viewer==0.0.19
striprtf==0.0.26
sympy==1.13.1
tenacity==8.5.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktoken==0.8.0
tokenizers==0.20.1
toml==0.10.2
tomli==2.0.2
tomlkit==0.13.2
torch==2.5.0
tornado==6.4.1
tqdm==4.66.5
transformers==4.46.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.32.0
w3lib==2.2.1
watchdog==5.0.3
websockets==13.1
wheel==0.44.0
wrapt==1.16.0
yarl==1.16.0